// console/src/generated/ChannelFieldDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChannelFieldControlDto } from "./ChannelFieldControlDto";

export type ChannelFieldDto = { key: string, i18n_key: string, control: ChannelFieldControlDto, required: boolean, advanced: boolean, default_value: string | null, options: Array<string>, };